+++ gatsby-config.js
@@ -27,6 +27,14 @@ module.exports = {
         lang: "en"
       }
     },
-    "gatsby-plugin-sitemap"
+    "gatsby-plugin-sitemap",
+    "gatsby-plugin-robots-txt",
+    {
+      resolve: "gatsby-plugin-google-analytics",
+      options: {
+        trackingId: "UA-142190723-1",
+        head: true
+      }
+    }
   ]
 };