+++ src/components/Gallery.js
@@ -19,9 +19,7 @@ class Gallery extends Component {
       selectedIndex
     }));
   }
-  renderGallery() {
-    const { images } = this.props;
-
+  renderGallery(images) {
     if (!images) return;
 
     const gallery = images.map((obj, i) => {
@@ -30,6 +28,7 @@ class Gallery extends Component {
           <a
             className="image fit thumb"
             href={obj.source}
+            title={obj.caption}
             onClick={e => {
               e.preventDefault();
               this.toggleLightbox(i);
@@ -52,7 +51,7 @@ class Gallery extends Component {
 
     return (
       <div>
-        {this.renderGallery()}
+        {this.renderGallery(images)}
         <ModalGateway>
           {lightboxIsOpen && (
             <Modal onClose={this.toggleLightbox}>