+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2019-06-16
+
+### Added [1.3.0]
+
+- Add Google Analytics
+- Add robots.txt
+- Add title property to a tags
+
+### Changed [1.3.0]
+
+- Update title and h1 tags
+- Pass images to render gallery function
+
+### Removed [1.3.0]
+
+- Remove yargs package
+
 ## [1.2.0] - 2019-06-15
 
 ### Changed [1.2.0]